Fix race condition in ExchangeContext.
Previously ExchangeContext relied on modifying a client's dialer just
before dialing, which was a race condition. This plumbs through contexts
in various paths so that ExchangeContext no longer has to rely on that
race condition. When a context with a deadline is used, and the client
also has a ReadTimeout or WriteTimeout configured, whichever is soonest
takes precedence. This allows us to implement the non-Context functions
in terms of the Context functions, by passing a context.Background().

Since this relies on tls.Dialer.DialContext, which is new in Go 1.15,
this should not be merged until Go 1.17 is released and the minimum
supported Go version is bumped.